+++ src/app/(user)/layout.tsx
@@ -19,7 +19,7 @@ export default async function UserLayout({ children }: { children: React.ReactNo
   return (
     <>
       <ConsentAwareGoogleAdSense
-        googleAdsenseId={process.env.NEXT_PUBLIC_GAS_ID ?? 'ca-pub-7412827340538951'}
+        googleAdsenseId={process.env.NEXT_PUBLIC_GAS_ID ?? ''}
       />
 
       <GlobalStructuredData />